# Use bigint-hash for hashing