--- a/package.json
+++ b/package.json
@@ -39,7 +39,7 @@
   "license": "MPL-2.0",
   "dependencies": {
     "async": "^1.4.2",
-    "keccak": "^1.4.0",
+    "bigint-hash": "^0.1.1",
     "level-ws": "0.0.0",
     "levelup": "^1.2.1",
     "memdown": "^1.0.0",
@@ -72,7 +72,7 @@
     "ts-loader": "^4.5.0",
     "ts-node": "^7.0.1",
     "typedoc": "^0.12.0",
-    "typescript": "~2.6.1",
+    "typescript": "github:calebsander/TypeScript#feature/bigint-lkg",
     "webpack": "^4.16.5",
     "webpack-cli": "^3.1.0"
   },