This PR switches from the keecak package to bigint-hash for hashing. This results in a 2x performance increase for hashing intensive functions, such as freshly calculating a root:

Before:
```
no-op: 198804±2.03% ops/s 0.005±0.0004 ms/op (60 runs)
put (empty tree): 172420±0.68% ops/s 0.006±0.0002 ms/op (80 runs)
get (empty tree): 137989±0.85% ops/s 0.007±0.0003 ms/op (71 runs)
generate 1k-10k-32-ran: 149±1.45% ops/s 6.719±0.4185 ms/op (71 runs)
generate 1k-1k-32-ran: 34.53±3.64% ops/s 28.962±4.4988 ms/op (70 runs)
generate 1k-1k-32-mir: 28.27±1.44% ops/s 35.376±2.1207 ms/op (67 runs)
generate 1k-9-32-ran: 12.76±1.32% ops/s 78.394±4.1429 ms/op (62 runs)
generate 1k-5-32-ran: 11.68±2.13% ops/s 85.588±7.0673 ms/op (58 runs)
generate 1k-3-32-ran: 10.88±1.70% ops/s 91.932±5.8518 ms/op (54 runs)
```

After:
```
no-op: 184645±2.05% ops/s 0.005±0.0005 ms/op (72 runs)
put (empty tree): 154683±1.50% ops/s 0.006±0.0004 ms/op (74 runs)
get (empty tree): 126926±2.51% ops/s 0.008±0.0009 ms/op (81 runs)
generate 1k-10k-32-ran: 700±2.64% ops/s 1.428±0.1674 ms/op (76 runs)
generate 1k-1k-32-ran: 66.53±7.66% ops/s 15.030±5.2846 ms/op (81 runs)
generate 1k-1k-32-mir: 61.69±1.38% ops/s 16.209±0.9778 ms/op (73 runs)
generate 1k-9-32-ran: 21.01±4.69% ops/s 47.607±8.3750 ms/op (54 runs)
generate 1k-5-32-ran: 20.23±4.25% ops/s 49.423±7.7204 ms/op (52 runs)
generate 1k-3-32-ran: 18.42±5.98% ops/s 54.278±11.4711 ms/op (48 runs)
```